pyproject.toml
[tool.poetry]
name = "codex-veritas"
version = "0.1.0"
description = "An AI software engineering platform built on the principle of Structural Honesty."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "codex", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
typer = {extras = ["rich"], version = "^0.12.3"}
tree-sitter = "0.21.0"
tree-sitter-languages = "1.9.1"
networkx = "^3.3"
chromadb = "^0.5.4"
sentence-transformers = "^3.0.1"
fastapi = "^0.111.1"
uvicorn = {extras = ["standard"], version = "^0.30.3"}
GitPython = "^3.1.43"
markdown2 = "^2.4.13"
python-dotenv = "^1.0.1"
google-generativeai = "^0.7.2"
# --- New Dependency for Sprint 8 ---
pygithub = "^2.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
black = "^24.4.2"
ruff = "^0.5.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

